Support Rack response bodies that allow to_ary (#22)
Newer Rack versions support bodies that respond to `to_ary` and in newer versions of Rails the Rails response body does as well. It seems previously we would not treat such response bodies correctly, meaning they would not get cached by Idempo (providing locking but no caching). This should be rectified with this PR - when running on Rack 3 (and newer Rails) `to_ary` will be called on the body, materializing it into an Array of Strings.